refactor: update import paths to use new wagmi configuration
