+++ templates/solidity-hardhat-wagmi/dapp-vue/src/components/Content.vue
@@ -2,7 +2,7 @@
 import { useAccount, useReadContract, useWaitForTransactionReceipt, useWriteContract } from '@wagmi/vue'
 import { computed, ref, watch } from 'vue'
 import { MESSAGE_BOARD_ADDRESS, MessageBoardABI } from '../config/contracts'
-import { passetHub } from '../config/reown'
+import { passetHub } from '../config/wagmi'
 import { shortenAddress } from '../utils/formatters'
 import Balance from './Balance.vue'
 import MessageCard from './MessageCard.vue'
