+++ templates/solidity-hardhat-wagmi/dapp-vue/src/config/wagmi.ts
